Rename User → Wallet; tighten error types
File and module rename: components/user{,/} → components/wallet{,/}.
Struct: User → Wallet. Error types renamed for idiomatic
module-relative naming: WalletError → Error,
BundleLoadError → BundleError. Dead variants pruned: Error::OutOfOrder,
BundleError::{ActionDecryptionFailed, InvalidActionIndex,
InvalidTransactionFormat} — none had constructors.
