+++ src/commands/test_issue_one.rs
@@ -11,7 +11,7 @@ use crate::commands::test_balances::{print_balances, TestBalances};
 use crate::components::db;
 use crate::components::rpc_client::reqwest::ReqwestRpcClient;
 use crate::components::transactions::{create_issue_transaction, mine, sync_from_height};
-use crate::components::user::User;
+use crate::components::wallet::Wallet;
 use crate::prelude::*;
 
 /// Run the simple issue test
@@ -25,7 +25,7 @@ impl Runnable for TestIssueOneCmd {
         let mut c = db::open();
         let mut rpc_client = ReqwestRpcClient::new(config.network.node_url());
         // Stable wallet identity so tree state and notes persist across runs.
-        let mut wallet = User::new(&mut c, &config.wallet.seed_phrase);
+        let mut wallet = Wallet::new(&mut c, &config.wallet.seed_phrase);
 
         let num_users = 1;
         let issuer_idx = 0;
