+++ src/commands/test_three_party.rs
@@ -19,7 +19,7 @@ use crate::commands::test_balances::{
 use crate::components::db;
 use crate::components::rpc_client::reqwest::ReqwestRpcClient;
 use crate::components::transactions::{create_issue_transaction, mine, sync_from_height};
-use crate::components::user::User;
+use crate::components::wallet::Wallet;
 use crate::prelude::*;
 
 /// Run the test scenario
@@ -33,7 +33,7 @@ impl Runnable for TestThreePartyCmd {
         let mut c = db::open();
         let mut rpc_client = ReqwestRpcClient::new(config.network.node_url());
         // Stable wallet identity so tree state and notes persist across runs.
-        let mut wallet = User::new(&mut c, &config.wallet.seed_phrase);
+        let mut wallet = Wallet::new(&mut c, &config.wallet.seed_phrase);
 
         sync_from_height(
             &mut c,
