+++ src/components.rs
@@ -5,6 +5,6 @@ pub mod miner;
 pub mod persistence;
 pub mod rpc_client;
 pub mod tree_state;
-pub mod user;
+pub mod wallet;
 
 pub mod transactions;
