+++ src/components/wallet/structs.rs
